Set up GraphQL Fixes #5
