+++ Cargo.toml
@@ -8,19 +8,24 @@ members = [".", "quasar_entities", "quasar_migrations", "stellar_node_entities"]
 
 [workspace.dependencies]
 sea-orm = "0.12.2"
+async-graphql = "6.0.7"
+async-graphql-axum = "6.0.7"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-quasar_entities = { path = "quasar_entities" }
-quasar_migrations = { path = "quasar_migrations" }
-stellar_node_entities = { path = "stellar_node_entities" }
+async-graphql = { workspace = true, features = ["playground"] }
+async-graphql-axum = { workspace = true }
+axum = "0.6"
 clap = { version = "4.4.2", features = ["derive"] }
 config = "0.13.3"
 log = "0.4.20"
 log4rs = "1.2.0"
+quasar_entities = { path = "quasar_entities" }
+quasar_migrations = { path = "quasar_migrations" }
 sea-orm = { workspace = true, features = ["sqlx-postgres", "runtime-tokio-rustls"] }
 serde = { version = "1.0.188", features = ["derive"] }
-tokio = { version = "1.32.0", features = ["full"] }
+stellar_node_entities = { path = "stellar_node_entities" }
 stellar-xdr = { version = "0.0.17", features = ["base64"] }
 thiserror = "1.0"
+tokio = { version = "1.32.0", features = ["full"] }
