+++ config/config.toml
@@ -2,3 +2,9 @@ quasar_database_url = "postgres://quasar:quasar@localhost:5432/quasar_developmen
 
 [ingestion]
 polling_interval = 5
+
+[api]
+host = "127.0.0.1"
+port = 8080
+depth_limit = 3
+complexity_limit = 16
