+++ quasar_entities/src/ledger.rs
@@ -1,11 +1,13 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.2
-
+use async_graphql::ComplexObject;
 use sea_orm::{entity::prelude::*, Set};
+
 use stellar_node_entities::ledgerheaders;
 use stellar_xdr::{Error, LedgerHeader, ReadXdr};
 
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, async_graphql::SimpleObject)]
 #[sea_orm(table_name = "ledgers")]
+#[graphql(complex)]
+#[graphql(name = "Ledgers")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
     pub hash: String,
@@ -25,6 +27,9 @@ pub struct Model {
 pub enum Relation {}
 
 impl ActiveModelBehavior for ActiveModel {}
+
+#[ComplexObject]
+impl Model {}
 
 impl TryFrom<ledgerheaders::Model> for ActiveModel {
     type Error = Error;
